eco_champions_backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
sqlalchemy>=2.0.36
psycopg[binary]==3.1.18
flask-migrate==4.0.7
alembic>=1.13.2
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.1
bcrypt==4.2.0
cloudinary==1.41.0
redis==5.0.8
eventlet==0.36.1
pytest==8.3.2
